man_analytics_app/requirements.txt
altair==5.3.0
annotated-types==0.6.0
astor==0.8.1
attrs==23.2.0
base58==2.1.1
bcrypt==4.1.2
blinker==1.8.1
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
extra-streamlit-components==0.1.71
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.34.0
streamlit-antd-components==0.3.2
streamlit-authenticator==0.3.2
streamlit-calendar==1.2.0
tenacity==8.2.3
thinc==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.1
wasabi==1.1.2
watchdog==4.0.0
wcag-contrast-ratio==0.9
weasel==0.3.4
